Use size_t instead of unsigned int Improve parallelization of ^ operator

+++ normalform/monomcoeff.h
@@ -7,13 +7,13 @@ namespace normalform {
 
 	using std::complex;
 
-	template<unsigned int N,class Tfloat> class CPolynom;
+	template<size_t N,class Tfloat> class CPolynom;
 
 	//friend function
-	template<unsigned int N,class Tfloat> class CMonomCoeff;
-	template<unsigned int N,class Tfloat> CMonomCoeff<N,Tfloat> operator*(CMonomCoeff<N,Tfloat>, const CMonomCoeff<N,Tfloat>&);
+	template<size_t N,class Tfloat> class CMonomCoeff;
+	template<size_t N,class Tfloat> CMonomCoeff<N,Tfloat> operator*(CMonomCoeff<N,Tfloat>, const CMonomCoeff<N,Tfloat>&);
 
-	template<unsigned int N,class Tfloat=double>
+	template<size_t N,class Tfloat=double>
 	class CMonomCoeff
 	{
 	public:
@@ -27,7 +27,7 @@ namespace normalform {
 		CMonomCoeff(const complex<Tfloat> c, const IntPower powers[2*N])
 		{
 			coeff = c;
-			for(unsigned int i=0; i<2*N; i++)
+			for(size_t i=0; i<2*N; i++)
 				monom[i] = powers[i];
 		};
 		CMonomCoeff(const typename CPolynom<N,Tfloat>::CMonomMap::const_iterator it)
@@ -39,7 +39,7 @@ namespace normalform {
 		friend CMonomCoeff<N,Tfloat> operator*<>(CMonomCoeff<N,Tfloat> lhs, const CMonomCoeff<N,Tfloat>& rhs);
 	};
 
-	template<unsigned int N,class Tfloat>
+	template<size_t N,class Tfloat>
 	inline CMonomCoeff<N,Tfloat> operator*(CMonomCoeff<N,Tfloat> lhs, const CMonomCoeff<N,Tfloat>& rhs)
 	{
 		lhs.monom *= rhs.monom;
